(refactor) Reorganize and rename components
