+++ src/components/modals/new-page.component.tsx
@@ -10,42 +10,39 @@ import {
   ModalHeader,
   TextInput,
 } from "@carbon/react";
-import { Page, Schema } from "../../../types";
-import { Edit } from "@carbon/react/icons";
+import { Page, Schema } from "../../types";
+import { Add } from "@carbon/react/icons";
 import { showToast } from "@openmrs/esm-framework";
+import styles from "./modals.scss";
 
-interface EditPageModalProps {
-  page: Page;
+interface NewPageModalProps {
+  pages: Array<Page>;
   schema: Schema;
   onSchemaUpdate: (schema: Schema) => void;
 }
 
-const EditPage: React.FC<EditPageModalProps> = ({
-  page,
+const CreatePage: React.FC<NewPageModalProps> = ({
+  pages,
   schema,
   onSchemaUpdate,
 }) => {
   const { t } = useTranslation();
-  const [openEditPageModal, setOpenEditPageModal] = useState(false);
+  const [openCreatePageModal, setOpenCreatePageModal] = useState(false);
   const [pageName, setPageName] = useState("");
-
-  useEffect(() => {
-    setPageName(page.label);
-  }, [page]);
 
   const handleSubmit = (event) => {
     event.preventDefault();
-    let pageName = event.target.pageName.value;
+    const newPage: Page = { label: pageName, sections: [] };
     try {
-      page.label = pageName == "" ? page.label : pageName;
+      pages.push(newPage);
       onSchemaUpdate({ ...schema });
       showToast({
         title: t("success", "Success!"),
         kind: "success",
         critical: true,
-        description: t("updatePage", "Page Updated"),
+        description: t("createPageSuccess", "Page Created"),
       });
-      setOpenEditPageModal(false);
+      setOpenCreatePageModal(false);
     } catch (error) {
       showToast({
         title: t("error", "Error"),
@@ -55,14 +52,19 @@ const EditPage: React.FC<EditPageModalProps> = ({
       });
     }
   };
+
+  useEffect(() => {
+    setPageName("");
+  }, [openCreatePageModal]);
+
   return (
     <>
       <div>
         <ComposedModal
-          open={openEditPageModal}
-          onClose={() => setOpenEditPageModal(false)}
+          open={openCreatePageModal}
+          onClose={() => setOpenCreatePageModal(false)}
         >
-          <ModalHeader title={t("editPage", "Edit Page")} />
+          <ModalHeader title={t("createPage", "Create Page")} />
           <Form onSubmit={handleSubmit}>
             <ModalBody>
               <FormGroup legendText={""}>
@@ -81,7 +83,7 @@ const EditPage: React.FC<EditPageModalProps> = ({
               </Button>
               <Button
                 kind={"secondary"}
-                onClick={() => setOpenEditPageModal(false)}
+                onClick={() => setOpenCreatePageModal(false)}
               >
                 {t("close", "Close")}
               </Button>
@@ -90,17 +92,17 @@ const EditPage: React.FC<EditPageModalProps> = ({
         </ComposedModal>
       </div>
       <Button
-        size="sm"
-        renderIcon={Edit}
-        iconDescription="Edit Page"
-        hasIconOnly
-        kind="ghost"
+        className={styles.CreatePageButton}
+        renderIcon={Add}
+        iconDescription="Create Page"
         onClick={() => {
-          setOpenEditPageModal(true);
+          setOpenCreatePageModal(true);
         }}
-      />
+      >
+        {t("createPage", "Create Page")}
+      </Button>
     </>
   );
 };
 
-export default EditPage;
+export default CreatePage;
