+++ src/components/modals/new-question.component.tsx
@@ -14,16 +14,10 @@ import {
   SelectItem,
   TextInput,
 } from "@carbon/react";
-import {
-  Answer,
-  Concept,
-  ConceptMapping,
-  Question,
-  Schema,
-} from "../../../types";
+import { Answer, Concept, ConceptMapping, Question, Schema } from "../../types";
 import { Add } from "@carbon/react/icons";
 import { showToast, useConfig } from "@openmrs/esm-framework";
-import { useConceptLookup } from "../../../hooks/useConceptLookup";
+import { useConceptLookup } from "../../hooks/useConceptLookup";
 import styles from "./modals.scss";
 
 interface CreateQuestionModalProps {
