+++ src/components/schema-editor/schema-editor.scss
